+++ CMakeLists.txt
@@ -274,7 +274,7 @@ add_library(
   src/s2geography/geoarrow.cc
   src/s2geography/geography_interface.cc
   src/s2geography/geography.cc
-  src/s2geography/geometry.cc
+  src/s2geography/tessellate.cc
   src/s2geography/linear-referencing.cc
   src/s2geography/op/cell.cc
   src/s2geography/op/point.cc
@@ -391,7 +391,7 @@ if(S2GEOGRAPHY_BUILD_TESTS)
   add_executable(distance_test src/s2geography/distance_test.cc)
   add_executable(geoarrow_test src/s2geography/geoarrow_test.cc)
   add_executable(geography_test src/s2geography/geography_test.cc)
-  add_executable(geometry_test src/s2geography/geometry_test.cc)
+  add_executable(tessellate_test src/s2geography/tessellate_test.cc)
   add_executable(linear_referencing_test
                  src/s2geography/linear-referencing_test.cc)
   add_executable(op_cell_test src/s2geography/op/cell_test.cc)
@@ -428,7 +428,7 @@ if(S2GEOGRAPHY_BUILD_TESTS)
   target_link_libraries(
     geography_test s2geography ${S2GEOGRAPHY_NANOARROW_TARGET}
     GTest::gtest_main GTest::gmock)
-  target_link_libraries(geometry_test s2geography GTest::gtest_main)
+  target_link_libraries(tessellate_test s2geography GTest::gtest_main)
   target_link_libraries(
     linear_referencing_test s2geography ${S2GEOGRAPHY_NANOARROW_TARGET}
     GTest::gtest_main GTest::gmock)
@@ -462,7 +462,7 @@ if(S2GEOGRAPHY_BUILD_TESTS)
   gtest_discover_tests(op_cell_test)
   gtest_discover_tests(predicates_test)
   gtest_discover_tests(geography_test)
-  gtest_discover_tests(geometry_test)
+  gtest_discover_tests(tessellate_test)
   gtest_discover_tests(wkt_writer_test)
   gtest_discover_tests(wkb_test)
   gtest_discover_tests(geoarrow_geography_test)
