+++ src/s2geography/tessellate_test.cc
@@ -1,4 +1,4 @@
-#include "s2geography/geometry.h"
+#include "s2geography/tessellate.h"
 
 #include <gtest/gtest.h>
